requirements.txt
featurecloud
bottle
numpy
scikit-learn
matplotlib
plotly
psutil
requests
kaleido
pandas
pyyaml
jsonpickle